question authorization using gate

+++ app/Http/Controllers/QuestionController.php
@@ -71,7 +71,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (\Gate::allows('update-question', $question)) {
+            return view('questions.edit', compact('question'));
+        }
+        abort(403, 'Access denied');
     }
 
     /**
@@ -83,11 +86,14 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        try {
-            $question->update($request->all());
-            return redirect()->route('questions.index')->with('success', 'Your question has been updated');
-        } catch (Exception $e) {
+        if (\Gate::allows('update-question', $question)) {
+            try {
+                $question->update($request->all());
+                return redirect()->route('questions.index')->with('success', 'Your question has been updated');
+            } catch (Exception $e) {
+            }
         }
+        abort(403, 'Access denied');
     }
 
     /**
@@ -98,10 +104,13 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
-        try {
-            $question->delete();
-            return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
-        } catch (Exception $e) {
+        if (\Gate::allows('delete-question', $question)) {
+            try {
+                $question->delete();
+                return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
+            } catch (Exception $e) {
+            }
         }
+        abort(403, 'Access denied');
     }
 }
